requirements.txt
# Kafka client for Python
kafka-python==2.0.2

# MongoDB client
pymongo==4.5.0

# JSON parsing
orjson==3.9.7

# Logging & utility libraries
loguru==0.7.2

# Data validation (optional, for strict data schemas)
pydantic==2.5.3

uvicorn
influxdb-client 
pandas 
scikit-learn 
joblib
fastapi